+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ### Improvements
 - [#1203](https://github.com/crypto-org-chain/chain-main/pull/1203) feat: use cosmos-sdk v0.53.4.
-- [#1205](https://github.com/crypto-org-chain/chain-main/pull/1205) Bump module version to v8 to follow semver convention
+- [#1205](https://github.com/crypto-org-chain/chain-main/pull/1205) Bump module version to v8 to follow semver 
+- [#1236](https://github.com/crypto-org-chain/chain-main/pull/1236) Add inflation decay
+
 ### Bugfixes
 - [#1209](https://github.com/crypto-org-chain/chain-main/pull/1209) Patch comet bft (GHSA-hrhf-2vcr-ghch)
-- [#1236](https://github.com/crypto-org-chain/chain-main/pull/1236) Add inflation decay
+
 
 
 *July 9, 2025*